feat: add ACL (#3)
* :arrow_up: add laravel-permission

* published files

* db seeders

* policies

* link policies with controllers

* update to handle roles

* update data seeders

* Add role to the tests

* new role composable

* filter nav items

* only admin places

* non admin pages

* lint

* role selector component

* add forms

* tests

* :arrow_up:

+++ backend/app/Http/Controllers/ReservationController.php
@@ -33,6 +33,8 @@ class ReservationController extends Controller
      */
     public function store(StoreReservationRequest $request, Event $event): Reservation
     {
+        $this->authorize('create', Reservation::class);
+
         // $event = EventDomain::tryToGetActiveEvent();
 
         // abort_unless($event, 422, 'Missing active event');
@@ -66,6 +68,8 @@ class ReservationController extends Controller
      */
     public function update(UpdateReservationRequest $request, Reservation $reservation): Reservation
     {
+        $this->authorize('update', Reservation::class);
+
         $event = $reservation->event()->first();
 
         EventDomain::updateReservationsAmounts($event);
@@ -85,6 +89,8 @@ class ReservationController extends Controller
      */
     public function destroy(Reservation $reservation): Reservation
     {
+        $this->authorize('delete', Reservation::class);
+
         $reservation->delete();
 
         return $reservation;
